changes required for kgrl to work and BARL to also work

+++ barl/acq/acquisition.py
@@ -863,7 +863,7 @@ class KGRLAcqFunction(AcqFunction):
         current_states = np.array(self.start_states)
         obs_dim = current_states.shape[-1]
         current_states = np.repeat(current_states[np.newaxis, :, :], self.params.num_fs, axis=0)
-        current_states = tf.convert_to_tensor(current_states, dtype=tf.float32)
+        current_states = tf.convert_to_tensor(current_states, dtype=tf.float64)
         model.initialize_function_sample_list(self.params.num_fs)
         f_batch_list = model.call_function_sample_list
         returns = 0
